+++ includes/settings/js/src/components/CreateContentModel.jsx
@@ -1,6 +1,7 @@
-import React, { useEffect, useContext, useState } from "react";
+/* global atlasContentModeler */
+import React, { useContext, useState } from "react";
 import { useForm } from "react-hook-form";
-import { Link, useHistory } from "react-router-dom";
+import { useHistory } from "react-router-dom";
 import { ModelsContext } from "../ModelsContext";
 import { insertSidebarMenuItem } from "../utils";
 import { useInputGenerator } from "../hooks";
@@ -9,6 +10,7 @@ import { showSuccess } from "../toasts";
 import Icon from "../../../../components/icons";
 import IconPicker from "./IconPicker";
 import { sprintf, __ } from "@wordpress/i18n";
+import { sendEvent } from "acm-analytics";
 
 const { apiFetch } = wp;
 
@@ -29,7 +31,6 @@ export default function CreateContentModel() {
 	const history = useHistory();
 	const [singularCount, setSingularCount] = useState(0);
 	const [pluralCount, setPluralCount] = useState(0);
-	const [icon, setIcon] = useState(0);
 	const [descriptionCount, setDescriptionCount] = useState(0);
 	const { dispatch } = useContext(ModelsContext);
 	const {
@@ -49,6 +50,11 @@ export default function CreateContentModel() {
 		})
 			.then((res) => {
 				if (res.success) {
+					sendEvent({
+						category: "Models",
+						action: "Model Created",
+					});
+
 					dispatch({ type: "addModel", data: res.model });
 					history.push(
 						atlasContentModeler.appPath +
@@ -75,6 +81,18 @@ export default function CreateContentModel() {
 				if (err.code === "acm_model_exists") {
 					setError("slug", {
 						type: "idExists",
+						message: err.message,
+					});
+				}
+				if (err.code === "acm_singular_label_exists") {
+					setError("singular", {
+						type: "exists",
+						message: err.message,
+					});
+				}
+				if (err.code === "acm_plural_label_exists") {
+					setError("plural", {
+						type: "exists",
 						message: err.message,
 					});
 				}
@@ -144,6 +162,15 @@ export default function CreateContentModel() {
 										</span>
 									</span>
 								)}
+							{errors.singular &&
+								errors.singular.type === "exists" && (
+									<span className="error">
+										<Icon type="error" />
+										<span role="alert">
+											{errors.singular.message}
+										</span>
+									</span>
+								)}
 							<span>&nbsp;</span>
 							<span className="count">{singularCount}/50</span>
 						</p>
@@ -195,6 +222,14 @@ export default function CreateContentModel() {
 										</span>
 									</span>
 								)}
+							{errors.plural && errors.plural.type === "exists" && (
+								<span className="error">
+									<Icon type="error" />
+									<span role="alert">
+										{errors.plural.message}
+									</span>
+								</span>
+							)}
 							<span>&nbsp;</span>
 							<span className="count">{pluralCount}/50</span>
 						</p>