+++ includes/settings/js/src/components/EditModelModal.jsx
@@ -5,31 +5,10 @@ import { ModelsContext } from "../ModelsContext";
 import Icon from "../../../../components/icons";
 import IconPicker from "./IconPicker";
 import { __ } from "@wordpress/i18n";
+import { sendEvent } from "acm-analytics";
 import { updateSidebarMenuItem } from "../utils";
+
 const { apiFetch } = wp;
-
-/**
- * Updates a model via the REST API.
- *
- * @param slug Model slug.
- * @param data Model data.
- */
-function updateModel(slug = "", data = {}) {
-	if (!slug.length || Object.keys(data).length === 0) {
-		return;
-	}
-
-	const updated = apiFetch({
-		path: `/wpe/atlas/content-model/${slug}`,
-		method: "PATCH",
-		_wpnonce: wpApiSettings.nonce,
-		data,
-	}).then((res) => {
-		return res;
-	});
-
-	return updated;
-}
 
 /**
  * The modal component for editing a content model.
@@ -55,6 +34,7 @@ export function EditModelModal({ model, isOpen, setIsOpen }) {
 		handleSubmit,
 		errors,
 		setValue,
+		setError,
 		formState: { isSubmitting },
 	} = useForm();
 
@@ -73,6 +53,46 @@ export function EditModelModal({ model, isOpen, setIsOpen }) {
 			boxSizing: "border-box",
 		},
 	};
+
+	/**
+	 * Updates a model via the REST API.
+	 *
+	 * @param slug Model slug.
+	 * @param data Model data.
+	 */
+	function updateModel(slug = "", data = {}) {
+		if (!slug.length || Object.keys(data).length === 0) {
+			return;
+		}
+
+		return apiFetch({
+			path: `/wpe/atlas/content-model/${slug}`,
+			method: "PATCH",
+			_wpnonce: wpApiSettings.nonce,
+			data,
+		})
+			.then((res) => {
+				sendEvent({
+					category: "Models",
+					action: "Model Updated",
+				});
+				return res;
+			})
+			.catch((err) => {
+				if (err.code === "acm_singular_label_exists") {
+					setError("singular", {
+						type: "exists",
+						message: err.message,
+					});
+				}
+				if (err.code === "acm_plural_label_exists") {
+					setError("plural", {
+						type: "exists",
+						message: err.message,
+					});
+				}
+			});
+	}
 
 	useEffect(() => {
 		Modal.setAppElement("#root");
@@ -96,10 +116,12 @@ export function EditModelModal({ model, isOpen, setIsOpen }) {
 			<form
 				onSubmit={handleSubmit(async (data) => {
 					const mergedData = { ...model, ...data };
-					updateSidebarMenuItem(model, data);
-					await updateModel(data.slug, mergedData);
-					dispatch({ type: "updateModel", data: mergedData });
-					setIsOpen(false);
+					const result = await updateModel(data.slug, mergedData);
+					if (result?.success) {
+						updateSidebarMenuItem(model, data);
+						dispatch({ type: "updateModel", data: mergedData });
+						setIsOpen(false);
+					}
 				})}
 			>
 				<div className="row">
@@ -149,6 +171,15 @@ export function EditModelModal({ model, isOpen, setIsOpen }) {
 												"Exceeds max length.",
 												"atlas-content-modeler"
 											)}
+										</span>
+									</span>
+								)}
+							{errors.singular &&
+								errors.singular.type === "exists" && (
+									<span className="error">
+										<Icon type="error" />
+										<span role="alert">
+											{errors.singular.message}
 										</span>
 									</span>
 								)}
@@ -206,6 +237,14 @@ export function EditModelModal({ model, isOpen, setIsOpen }) {
 										</span>
 									</span>
 								)}
+							{errors.plural && errors.plural.type === "exists" && (
+								<span className="error">
+									<Icon type="error" />
+									<span role="alert">
+										{errors.plural.message}
+									</span>
+								</span>
+							)}
 							<span>&nbsp;</span>
 							<span className="count">{pluralCount}/50</span>
 						</p>