+++ includes/settings/js/src/reducer.js
@@ -72,15 +72,12 @@ export function reducer(state, action) {
 				editing: false,
 			};
 			return { ...state };
-		case "setTitleField":
 		case "setFeaturedImageField":
 			const fields = state[action.model]["fields"];
-			const setType =
-				action.type === "setTitleField" ? "isTitle" : "isFeatured";
 			Object.values(fields).forEach((field) => {
-				state[action.model]["fields"][field.id][setType] = false;
+				state[action.model]["fields"][field.id]["isFeatured"] = false;
 			});
-			state[action.model]["fields"][action.id][setType] = true;
+			state[action.model]["fields"][action.id]["isFeatured"] = true;
 			return { ...state };
 		case "setFieldProperties":
 			action.properties.forEach((property) => {