+++ tests/acceptance/CreateContentModelMultipleChoiceFieldCest.php
@@ -2,13 +2,16 @@
 
 class CreateContentModelMultipleChoiceFieldCest {
 
+	public function _before( \AcceptanceTester $i ) {
+		$i->loginAsAdmin();
+		$content_model = $i->haveContentModel( 'Candy', 'Candies' );
+		$i->amOnWPEngineEditContentModelPage( $content_model['slug'] );
+	}
+
 	/**
 	 * Ensure a user can add a multile choice field from a model and remove it.
 	 */
 	public function i_can_add_a_multiple_choice_field_to_a_content_model( AcceptanceTester $i ): void {
-		$i->loginAsAdmin();
-		$i->haveContentModel( 'Candy', 'Candies' );
-		$i->wait( 1 );
 		$i->click( 'Multiple Choice (Beta)', '.field-buttons' );
 		$i->wait( 1 );
 		$i->fillField( [ 'name' => 'name' ], 'Favorite Animal' );
@@ -36,9 +39,6 @@ class CreateContentModelMultipleChoiceFieldCest {
 	 * Ensure a user cannot create two options with the same key value api identifier.
 	 */
 	public function i_cannot_add_a_duplicate_identifier_name_for_two_choices( AcceptanceTester $i ): void {
-		$i->loginAsAdmin();
-		$i->haveContentModel( 'Candy', 'Candies' );
-		$i->wait( 1 );
 		$i->click( 'Multiple Choice (Beta)', '.field-buttons' );
 		$i->wait( 1 );
 		$i->fillField( [ 'name' => 'name' ], 'Favorite Animal' );
@@ -60,9 +60,6 @@ class CreateContentModelMultipleChoiceFieldCest {
 	 * Ensure a user cannot save a choice with no name or slug.
 	 */
 	public function i_cannot_save_a_blank_choice( AcceptanceTester $i ): void {
-		$i->loginAsAdmin();
-		$i->haveContentModel( 'Candy', 'Candies' );
-		$i->wait( 1 );
 		$i->click( 'Multiple Choice (Beta)', '.field-buttons' );
 		$i->wait( 1 );
 		$i->fillField( [ 'name' => 'name' ], 'Favorite Animal' );
@@ -82,9 +79,6 @@ class CreateContentModelMultipleChoiceFieldCest {
 	 * Ensure a user cannot create two choices with the same name.
 	 */
 	public function i_cannot_add_a_duplicate_name_for_two_choices( AcceptanceTester $i ): void {
-		$i->loginAsAdmin();
-		$i->haveContentModel( 'Candy', 'Candies' );
-		$i->wait( 1 );
 		$i->click( 'Multiple Choice (Beta)', '.field-buttons' );
 		$i->wait( 1 );
 		$i->fillField( [ 'name' => 'name' ], 'Favorite Animal' );