+++ tests/acceptance/CreateFieldWithIsTitleCest.php
@@ -1,5 +1,7 @@
 <?php
 
+use Codeception\Util\Locator;
+
 class CreateFieldWithIsTitleCest {
 
 	/**
@@ -7,8 +9,8 @@ class CreateFieldWithIsTitleCest {
 	 */
 	public function i_can_create_a_content_model_text_field_with_is_title( AcceptanceTester $i ) {
 		$i->loginAsAdmin();
-		$i->haveContentModel( 'Candy', 'Candies' );
-		$i->wait( 1 );
+		$content_model = $i->haveContentModel( 'Candy', 'Candies' );
+		$i->amOnWPEngineEditContentModelPage( $content_model['slug'] );
 
 		$i->click( 'Text', '.field-buttons' );
 		$i->wait( 1 );
@@ -21,5 +23,16 @@ class CreateFieldWithIsTitleCest {
 		$i->see( 'Text', '.field-list div.type' );
 		$i->see( 'Name', '.field-list div.widest' );
 		$i->see( 'entry title', '.field-list div.tags' );
+	}
+
+	public function i_can_not_create_a_second_text_field_with_is_title( AcceptanceTester $i ) {
+		$this->i_can_create_a_content_model_text_field_with_is_title( $i );
+
+		// Add a second text field.
+		$i->click( Locator::lastElement( '.add-item' ) );
+		$i->click( 'Text', '.field-buttons' );
+
+		// Should not be possible to set another text field as the title.
+		$i->dontSee( 'Use this field as the entry title', '.field-form' );
 	}
 }