+++ tests/acceptance/EditContentModelCest.php
@@ -1,19 +1,23 @@
 <?php
+
 class EditContentModelCest {
-	public function i_can_update_an_existing_content_model( AcceptanceTester $i ) {
-		$i->resizeWindow( 1024, 1024 );
+
+	public function _before( \AcceptanceTester $i ) {
+		$i->maximizeWindow();
 		$i->loginAsAdmin();
 
 		// First, create a new model.
-		$i->haveContentModel( 'Candy', 'Candies' );
-		$i->wait( 1 );
+		$content_model = $i->haveContentModel( 'Candy', 'Candies' );
+		$i->amOnWPEngineEditContentModelPage( $content_model['slug'] );
 
 		// Invoke edit mode.
 		$i->amOnWPEngineContentModelPage();
 		$i->click( '.model-list button.options' );
 		$i->click( '.dropdown-content a.edit' );
 		$i->see( 'Edit Candies' );
+	}
 
+	public function i_can_update_an_existing_content_model( AcceptanceTester $i ) {
 		// Update the model data.
 		$i->fillField( [ 'name' => 'singular' ], 'Cat' );
 		$i->fillField( [ 'name' => 'plural' ], 'Cats' );
@@ -40,4 +44,22 @@ class EditContentModelCest {
 		$menu_label = $i->grabTextFrom( '#menu-posts-candy .wp-menu-name' );
 		$i->assertEquals( 'Cats', $menu_label );
 	}
+
+	public function i_see_a_warning_if_the_model_singular_name_is_reserved( AcceptanceTester $i ) {
+		// Update the model data.
+		$i->fillField( [ 'name' => 'singular' ], 'Post' ); // 'post' is in use.
+		$i->click( 'Save' );
+		$i->wait( 1 );
+
+		$i->see( 'singular name is in use', '.ReactModal__Content' );
+	}
+
+	public function i_see_a_warning_if_the_model_plural_name_is_reserved( AcceptanceTester $i ) {
+		$i->fillField( [ 'name' => 'plural' ], 'Posts' ); // 'posts' is in use.
+		$i->click( 'Save' );
+		$i->wait( 1 );
+
+		$i->see( 'plural name is in use', '.ReactModal__Content' );
+	}
+
 }